+++ src/lib/utils/read_meta_stack.ts
